+++ hive-event-listeners/apiary-gluesync-listener/src/main/java/com/expediagroup/apiary/extensions/gluesync/listener/ApiaryGlueSync.java
@@ -38,6 +38,8 @@ import com.amazonaws.services.glue.AWSGlueClientBuilder;
 import com.amazonaws.services.glue.model.AlreadyExistsException;
 import com.amazonaws.services.glue.model.EntityNotFoundException;
 
+import com.expediagroup.apiary.extensions.gluesync.listener.metrics.MetricConstants;
+import com.expediagroup.apiary.extensions.gluesync.listener.metrics.MetricService;
 import com.expediagroup.apiary.extensions.gluesync.listener.service.GlueDatabaseService;
 import com.expediagroup.apiary.extensions.gluesync.listener.service.GluePartitionService;
 import com.expediagroup.apiary.extensions.gluesync.listener.service.GlueTableService;
@@ -46,29 +48,34 @@ import com.expediagroup.apiary.extensions.gluesync.listener.service.IsIcebergTab
 public class ApiaryGlueSync extends MetaStoreEventListener {
 
   private static final Logger log = LoggerFactory.getLogger(ApiaryGlueSync.class);
-  private final IsIcebergTablePredicate isIcebergPredicate = new IsIcebergTablePredicate();
 
   private final AWSGlue glueClient;
   private final GlueDatabaseService glueDatabaseService;
   private final GlueTableService glueTableService;
   private final GluePartitionService gluePartitionService;
+  private final IsIcebergTablePredicate isIcebergPredicate;
+  private final MetricService metricService;
 
   public ApiaryGlueSync(Configuration config) {
     super(config);
-    glueClient = AWSGlueClientBuilder.standard().withRegion(System.getenv("AWS_REGION")).build();
+    this.glueClient = AWSGlueClientBuilder.standard().withRegion(System.getenv("AWS_REGION")).build();
     String gluePrefix = System.getenv("GLUE_PREFIX");
-    glueDatabaseService = new GlueDatabaseService(glueClient, gluePrefix);
-    glueTableService = new GlueTableService(glueClient, gluePrefix);
-    gluePartitionService = new GluePartitionService(glueClient, gluePrefix);
+    this.glueDatabaseService = new GlueDatabaseService(glueClient, gluePrefix);
+    this.glueTableService = new GlueTableService(glueClient, gluePrefix);
+    this.gluePartitionService = new GluePartitionService(glueClient, gluePrefix);
+    this.isIcebergPredicate = new IsIcebergTablePredicate();
+    this.metricService = new MetricService();
     log.debug("ApiaryGlueSync created");
   }
 
-  public ApiaryGlueSync(Configuration config, AWSGlue glueClient, String gluePrefix) {
+  public ApiaryGlueSync(Configuration config, AWSGlue glueClient, String gluePrefix, MetricService metricService) {
     super(config);
     this.glueClient = glueClient;
-    glueDatabaseService = new GlueDatabaseService(glueClient, gluePrefix);
-    glueTableService = new GlueTableService(glueClient, gluePrefix);
-    gluePartitionService = new GluePartitionService(glueClient, gluePrefix);
+    this.glueDatabaseService = new GlueDatabaseService(glueClient, gluePrefix);
+    this.glueTableService = new GlueTableService(glueClient, gluePrefix);
+    this.gluePartitionService = new GluePartitionService(glueClient, gluePrefix);
+    this.isIcebergPredicate = new IsIcebergTablePredicate();
+    this.metricService = metricService;
     log.debug("ApiaryGlueSync created");
   }
 
@@ -80,9 +87,14 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
     Database database = event.getDatabase();
     try {
       glueDatabaseService.create(database);
+      metricService.incrementCounter(MetricConstants.LISTENER_DATABASE_SUCCESS);
     } catch (AlreadyExistsException e) {
       log.info(database + " database already exists in glue, updating....");
       glueDatabaseService.update(database);
+      metricService.incrementCounter(MetricConstants.LISTENER_DATABASE_SUCCESS);
+    } catch (Exception e) {
+      log.error("Failed create database {} in glue", database.getName(), e);
+      metricService.incrementCounter(MetricConstants.LISTENER_DATABASE_FAILURE);
     }
   }
 
@@ -92,7 +104,13 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
       return;
     }
     Database database = event.getDatabase();
-    glueDatabaseService.delete(database);
+    try {
+      glueDatabaseService.delete(database);
+      metricService.incrementCounter(MetricConstants.LISTENER_DATABASE_SUCCESS);
+    } catch (Exception e) {
+      log.error("Failed drop database {} in glue", database.getName(), e);
+      metricService.incrementCounter(MetricConstants.LISTENER_DATABASE_FAILURE);
+    }
   }
 
   @Override
@@ -103,9 +121,14 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
     Table table = event.getTable();
     try {
       glueTableService.create(table);
+      metricService.incrementCounter(MetricConstants.LISTENER_TABLE_SUCCESS);
     } catch (AlreadyExistsException e) {
       log.info(table + " table already exists in glue, updating....");
       glueTableService.update(table);
+      metricService.incrementCounter(MetricConstants.LISTENER_TABLE_SUCCESS);
+    } catch (Exception e) {
+      log.error("Failed create table {}.{} in glue", table.getDbName(), table.getTableName(), e);
+      metricService.incrementCounter(MetricConstants.LISTENER_TABLE_FAILURE);
     }
   }
 
@@ -117,8 +140,12 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
     Table table = event.getTable();
     try {
       glueTableService.delete(table);
+      metricService.incrementCounter(MetricConstants.LISTENER_TABLE_SUCCESS);
     } catch (EntityNotFoundException e) {
       log.info(table + " table doesn't exist in glue catalog");
+    } catch (Exception e) {
+      log.error("Failed drop table {}.{} in glue", table.getDbName(), table.getTableName(), e);
+      metricService.incrementCounter(MetricConstants.LISTENER_TABLE_FAILURE);
     }
   }
 
@@ -136,9 +163,14 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
         return;
       }
       glueTableService.update(newTable);
+      metricService.incrementCounter(MetricConstants.LISTENER_TABLE_SUCCESS);
     } catch (EntityNotFoundException e) {
       log.info(newTable + " table doesn't exist in glue, creating....");
       glueTableService.create(newTable);
+      metricService.incrementCounter(MetricConstants.LISTENER_TABLE_SUCCESS);
+    } catch (Exception e) {
+      log.error("Failed alter table {}.{} in glue", oldTable.getDbName(), oldTable.getTableName(), e);
+      metricService.incrementCounter(MetricConstants.LISTENER_TABLE_FAILURE);
     }
   }
 
@@ -148,6 +180,7 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
     glueTableService.create(newTable);
     glueTableService.copyPartitions(newTable, glueTableService.getPartitions(oldTable));
     glueTableService.delete(oldTable);
+    metricService.incrementCounter(MetricConstants.LISTENER_TABLE_SUCCESS);
     long duration = System.currentTimeMillis() - startTime;
     log.info("{} glue table rename to {} finised in {}ms", oldTable.getTableName(), newTable.getTableName(), duration);
   }
@@ -167,8 +200,13 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
       Partition partition = partitions.next();
       try {
         gluePartitionService.create(table, partition);
+        metricService.incrementCounter(MetricConstants.LISTENER_PARTITION_SUCCESS);
       } catch (AlreadyExistsException e) {
         gluePartitionService.update(table, partition);
+        metricService.incrementCounter(MetricConstants.LISTENER_PARTITION_SUCCESS);
+      } catch (Exception e) {
+        log.error("Failed add partition on table {}.{} in glue", table.getDbName(), table.getTableName(), e);
+        metricService.incrementCounter(MetricConstants.LISTENER_PARTITION_FAILURE);
       }
     }
   }
@@ -184,7 +222,11 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
       Partition partition = partitions.next();
       try {
         gluePartitionService.delete(table, partition);
-      } catch (EntityNotFoundException ignored) {}
+        metricService.incrementCounter(MetricConstants.LISTENER_PARTITION_SUCCESS);
+      } catch (Exception e) {
+        log.error("Failed drop partition on table {}.{} in glue", table.getDbName(), table.getTableName(), e);
+        metricService.incrementCounter(MetricConstants.LISTENER_PARTITION_FAILURE);
+      }
     }
   }
 
@@ -197,8 +239,13 @@ public class ApiaryGlueSync extends MetaStoreEventListener {
     Partition partition = event.getNewPartition();
     try {
       gluePartitionService.update(table, partition);
+      metricService.incrementCounter(MetricConstants.LISTENER_PARTITION_SUCCESS);
     } catch (EntityNotFoundException e) {
       gluePartitionService.create(table, partition);
+      metricService.incrementCounter(MetricConstants.LISTENER_PARTITION_SUCCESS);
+    } catch (Exception e) {
+      log.error("Failed alter partition on table {}.{} in glue", table.getDbName(), table.getTableName(), e);
+      metricService.incrementCounter(MetricConstants.LISTENER_PARTITION_FAILURE);
     }
   }
 }